// src/enums/apiKey.ts
export const ACCESS_TOKEN = localStorage.getItem('access_token')?localStorage.getItem('access_token'):''
export enum ApiKey {
    CLIENT_ID = '2356',
    CLIENT_SECRET = 'v3.r.137440105.ffdbab114f92b821eac4e21f485343924a773131.06c3bdbb8446aeb91c35b80c42ff69eb9c457948',
    LOGIN = 'sergei.stralenia@gmail.com',
    PASSWORD = 'paralect123',
}